+++ CHANGES.rst
@@ -8,7 +8,7 @@ Changes to be released in next version
  * MXCallKitAdapter: Update incoming calls if answered from application UI.
  * MXFileStore: Logs all files when a data corruption is detected (to track vector-im/element-ios/issues/4921).
  * MXCallManager: Fix call transfers flow for all types of transfers.
- * VoIP: Implement asserted identity for calls.
+ * VoIP: Implement asserted identity for calls: MSC3086 (matrix-org/matrix-doc/pull/3086).
 
 🐛 Bugfix
  * MXTools: Fix bad linkification of matrix alias and URL (vector-im/element-ios/issues/4258).